fix loading of existing model from disk
When the model is exported and written to disk it is stored as a JSON string.
When read back in using json.load, it is stored as a dict.
markovify.Text.from_json expects a JSON string, so it was throwing a TypeError (: the JSON object must be str, not 'dict').
No idea why this only broke now since haven't touched any of the code...